Add + initialize image container in GameActivity
It's called `imgs` and it's a GridLayout that will eventually contain
ImageViews. It's initialized in GameActivity.onCreate() with a rowCount
and columnCount generated from the `difficulty` variable. Currently
`difficulty` is just a variable but it'll eventually be passed in the
intent from the main menu.

There are 3 values for `difficulty`: 4 (easy), 6 (normal), and 9 (hard).
The difficulty numbers are the number of ImageViews in `imgs` at each
difficulty. Note that if anything other than 4, 6, or 9 is passed in the
intent (eventually), an error will be logged and the GridLayout will be
set to 0x0, which will *probably* cause a crash.

One more change: the root layout in activity_game.xml was changed to
RelativeLayout.