Add clipart in assets tree: apple + beach ball
The clipart is for testing purposes but will make it into the final
build. In the assets tree, there are 2 directories (normals and
variants), each containing the same amount of 256x256 PNG files named
the same. The file in the 'normals' directory is the normal clipart (the
one which will cause a loss when clicked), and the one in the 'variants'
directory is the variant clipart (the one that will earn points when
clicked). Currently there's an apple (apple.png) and a beach ball
(beach-ball.png) for testing purposes.