Add ImageSupplier class with loading capability
ImageSupplier is a package-private class that is responsible for loading
and storing the normal/variant image pairs and giving a random pair when
requested. It stores the pairs as Pair<Drawable, Drawable>, with the
normal first. It currently loads all the images as Drawables in a load()
method, to which the constructor delegates. Note that this means that
if there are a lot of images, a significant amount of memory may be used
up by the Drawables; however, it does speed up the displaying of the
images (I think; none of this has been tested).

Currently it can only load images in the constructor; eventually it'll
have a method to return a random image pair.