Add click detection on variant/normal images
Each ImageSwitcher now has an ID corresponding to their index in imgs.
There is now a variantId field holding the ID/index of the variant
ImageSwitcher.

GameActivity is now a View.OnClickListener and handles clicks for the
ImageSwitchers. The onClick() method uses the variantId field to
determine if it is handling a click on the variant or not, and then
delegates to either onVariantClick() or onNormalClick().

onVariantClick() will eventually add score, reset the timer, etc; it
currently only calls updateImages(). onNormalClick() will probably
eventually be renamed onLose() and will stop the timer, wait, etc.

Ayy! Now it's somewhat playable!