Add animations when updating images
There are now in/out animations for the ImageSwitchers. The image
simply zooms in or out linearly in 200ms. The animation is defined in
res/anim/grow_in.xml; since the out animation is an exact reverse of the
in animation, a new ReverseInterpolator is used to reverse it. The
ReverseInterpolator reverses the interpolation of whatever interpolator
it's passed; in this case, whatever interpolator is defined in res/anim/
grow_in.xml, which is (currently) LinearInterpolator