Add countdown circle, rename GameActivity methods
Add http://github.com/eralpyucel/CircleProgressView as a dependency via
jitpack.io. Add a CircleProgressView, @id/countdown_circle, at the top
of the screen. This circle will eventually count down the time, as well
as show the current score for the round.

Add dimens.xml, which holds dimen values for the countdown circle and
will eventually hold all the others as well.

In GameActivity:
- rename onVariantClick() -> nextRound()
- rename onNormalClick() -> onLose()
- call nextRound() instead of updateImages() (except in nextRound())
- change up the order of methods