Add animation to the countdown circle
The countdown circle now actually counts down. It uses a ValueAnimator
partially defined in res/animator/countdown.xml to do this. It's linear.

The duration of the animation, as well as the starting number appearing
inside the circle, are determined by a function of the round number and
the parameters defined in the new file res/values/integers.xml.

The duration of the animation (the amount of time the user has to click
the variant) is defined by the following relation:

    d = a*cos(c*r*pi) + a + b if r > min_time_rounds else min_time_ms

where d is the duration, r is the round number, and:
- a = (initial_time_ms - min_time_ms) / 2
- b = min_time_ms
- c = 1 / min_time_rounds

The max score (the maximum possible score the user could achieve in any
given round) is given by the following linear relation:

    s = ar + b

where s is the max score, r is the round number, a is the
score_per_round integer defined in integers.xml (currently set to 10),
and b is the min_score variable defined in integers.xml (currently set
to 100).

Sorry for the crappy commit message, it's late.