Add score step to prevent unreadable round scores
Previously, the round score (the number in the middle of the countdown
circle) decreased by 1; it was *very* fast and nigh unreadable. Now it
decreases in multiples of @integer/score_step (currently set to 10). So
it decreases e.g. 100->90->80->etc instead of 100->99->98->etc. This
makes the number much more readable.

Also scoreForRound is a field in GameActivity now.