Add score
There's now a big purple TextView, @id/score_text, at the bottom of the
screen. It updates with the scorePerRound at the end of every round.
It's got a smallish label above it saying Score (@string/score_label).

There's a special method for updating score_text, addToScore(). This
method takes an int to add to the score, adds it, and then updates
score_text. Eventually maybe there'll be an animation to go along with
the addition.

Also score is an int field in GameActivity now.