Add best/average score tracking with shared prefs
Add "bestScore", "totalScore", and "timesPlayed" SharedPreferences,
which are used and updated by PostGameActivity and used for the best
and average score displays. It's totally insecure -- the root user can
change the shared preferences to whatever they want because it isn't
encrypted or anything -- but that doesn't matter because whenever Google
Play Games is eventually integrated, it won't use the values stored in
SharedPreferences.