Add difficulty button animation - sliding up
In MainActivity, when the play button is pressed, the difficulty
buttons now slide up from the play button into their positions. This is
accomplished through intializing the animations ahead of time, in a list
in MainActivity (slideUpAnims). The difficulty buttons are also now
initialized into a list (difficultyBtns). The slide up animations are
TranslateAnimations and have a variable fromYDelta and duration.

The duration calculation uses the new @integer/slide_up_speed, which is
the amount of time in milliseconds that a difficulty button will take to
move 100dp (pre-interpolation). The interpolator for the slide up
animations is DecelerateInterpolator.