Make difficulty buttons functional
The difficulty buttons now call onEasyClick()/onNormalClick()/
onHardClick() in MainActivity, all of which delegate to a new private
method, MainActivity.toGameActivity(), which takes a difficulty (4/6/9)
as a parameter and goes to the game activity with that difficulty as the
"DIFFICULTY" extra.

GameActivity now takes its difficulty from the intent's "DIFFICULTY"
extra; it's no longer hardcoded to be 4 (easy).

PostGameActivity now sets the "DIFFICULTY" extra in the intent -- but
there aren't any difficulty buttons in PostGameActivity (yet) so it's
hardcoded to be 4.