Abstract buttons into DifficultyButtonsFragment
The difficulty buttons and their logic is now encapsulated in
DifficultyButtonsFragment, which can be dropped into any Activity in
which difficulty buttons are needed. The difficulty buttons logic has
entirely been moved from MainActivity to DifficultyButtonsFragment. The
fragment is currently in MainActivity and will eventually be in
PostGameActivity as well.

Since different views have different strings that the play button should
say, DifficultyButtonsFragment has two parameters, show_text and
hide_text, for the strings that should appear on the play button.
show_text is shown when clicking on the play button would cause the
buttons to appear, and hide_text is shown when doing so would cause them
to disappear. The fragment takes these arguments in two ways:
programmatically, through the newInstance() static methods or through
settings the bundle of arguments another way, or through XML, by setting
the custom attributes app:show_text and app:hide_text. (The attributes
use snake_case instead of camelCase because apparently showText is
reserved). Note that hide_text defaults to @string/close_difficulty_btns
but show_text must always be specified.