Add difficulty label on PostGameActivity
It's displayed above the stats and says "[difficulty] Mode", e.g. "Easy
Mode" in @color/colorPrimary. Getting the difficulty to PostGameActivity
required saving the difficulty to an instance variable in GameActivity,
then making a "DIFFICULTY" extra a necessary part of the Intent to
PostGameActivity.

Also: solved the stats being weirdly low on the screen.