Add invisible (for now) pause overlay
This overlay is a separate layout file (game_paused.xml) that is
<include>'d into activity_game.xml. It's currently invisible, and will
become visible when the (nonexistent, yet) pause button is pressed.

It consists of merely a TextView with the text "Paused" and a button
entitled "Resume Game". Eventually, the countdown circle, score text
view, and score label will have their colours changed and will be
brought in front of the pause overlay, to form a contiguous UI.

There is one new dimension that is currently unused,
@dimen/unpause_countdown_text_size. When the "Resume Game" button is
pressed in the pause overlay, the paused TextView (@id/paused_text) will
be set to this size and its text will go 3 -> 2 -> 1, at which point
the pause overlay will disappear and the game will unpause.