Add pause button with custom drawable
There is now an XML drawable, pause.xml, that's a pause button. It's
(in vexillological terms) a vertical tricolour of black, transparency,
and black on a 1:1 square field.

It's the drawable for a 25x25dp ImageButton in GameActivity positioned
at the top-right of the screen. Eventually that button will bring up the
pause overlay and pause the game.