Add functionality to pause button
It now pauses all the animations if they're running (except for the
ImageSwitcher in/out animations, but hopefully that's not a problem),
then brings up the pause overlay, brings the countdown circle, score
text, and the score label to the front, and changes their colour scheme
to work better on the dark background. For the changed colour scheme (as
well as easier colour management in general) there are new colours:
@color/gameMain, set to @color/colorAccent and used for the text in the
game; @color/gamePaused, set to #e040fb and used for the text when
paused; @color/countdownCircle(Finished|Unfinished), used for the
finished/unfinished stroke colours for the countdown circle in-game; and
@color/countdownCirclePaused(Finished|Unfinished), used for the finished
/unfinished stroke colours for the countdown circle when paused.

You can't get out of the pause screen yet, there's no unpause
functionality. That'll be later.