Add @id/unpauseCountdownText, for the countdown
This TextView is for the large countdown text numbers which appear in
the centre of the screen. Its job was previously done by @id/pausedText,
which moved and grew, but that was far too complicated so another
TextView was introduced instead. Now the pausing and unpausing logic is
far simpler.