Add quit button to the pause overlay
There is now a "button" (it's got a transparent background, so really
it's just a clickable TextView) called @id/quit_button. When clicked, a
confirmation AlertDialog (with the Theme.Holo.Dialog theme) appears with
the message "Are you sure you want to quit the game?" (@string/
quit_message) and the two options "Yes, quit" (@string/quit_action) and
"Cancel". Choosing "Yes, quit" will finish() the GameActivity, which
brings the user back to the previous thing on the stack, MainActivity.
Meanwhile, choosing "Cancel" simply dismisses the dialog.

The button is horizontally centered and appears below the resume button.
It has no background and a "subtle" colour (@color/textSubtleOnDark). It
also disappears and reappears when the resume button and the paused text
does.