Add throwing image functionality in MainActivity
MainActivity.throwImage() is now a method. When called, it constructs an
ImageView with a randomly chosen image (normal or variant), generates a
random parabolic path for the image to follow (thus giving it the
appearance of being thrown), and also generates a random number of times
for the image to rotate. It then puts all this into an AnimatorSet and
plays it, removing the ImageView from its parent afterwards.

The equation of a parabola in vertex form is the following:
    y = a(x - h)^2 + k
where (h, k) is the vertex of the parabola. To generate a random
parabola, we randomly generate h and k (called minBias and
vertexHorizBias respectively) within ranges. We also generate a starting
position for the image, or a y-intercept (variable: b). With this, we
have enough information to solve for a:
    a = (b - k) / h^2
We now have all the information we need for the parabola. The image
follows this parabolic path.

Rotation is done through generating a random starting orientation for
the image, then generating a random number of times that the image will
rotate. The rotation animation also either accelerates or decelerates;
which one is chosen at random.

Since the parent layout is a ConstraintLayout, moving the image is done
through manipulating the horizontal and vertical biases. Thus, all the
translation is on a scale from 0f to 1f.

The ranges are all in floats and are stored in dimens.xml using a hack
described here: https://stackoverflow.com/a/8780360.

The ranges are:
- vertical bias (for starting position and vertex minimum):
  @dimen/[min|max]_thrown_image_vertical_bias (currently 0.3-0.7)
- horizontal maximum at the vertex:
  @dimen/[min|max]_thrown_image_horizontal_bias (currently 0.6-1.2)
  Note that if the vertex horizontal bias is over 1 it won't be shown.
- number of times the image will rotate:
  @dimen/[min|max]_thrown_image_rotate_times (currently 0.3-1.8)

Other resources:
- @dimen/min_start_min_thrown_image_bias_diff - the minimum difference
  between the starting and minimum vertical biases. Currently 0.2.
- @dimen/thrown_image_size - the width and height of thrown images.
  Currently 40dp.
- @integer/image_throw_duration - the duration of the throwing
  animation, in ms. Currently 750.

Also add Util.getFloatResource() and Util.randomFloatBetween(). The
former is a utility for getting a float resource, using the hack
described in the above StackOverflow link. The later is a utility for
getting a random float between two values; abstracting this makes
MainActivity.throwImage() less confusing.

Note that throwImage() currently isn't called. It will be in a
subsequent commit.