Add ButtonStyle, apply it to all buttons
Its parent is Widget.AppCompat.Button.Borderless and it's defined in
styles.xml. It holds many of the common attributes of all the buttons in
the app. All buttons now have that style.

Note that the style includes android:margin[Left|Right], which was being
set directly on the difficulty buttons fragment. It's now set on the
individual buttons of the fragment through the style.