Fix crashes when recreating fragment activities
The activities containing DifficultyButtonsFragment (MainActivity and
PostGameActivity) would crash whenever they were destroyed + recreated.
This was because DifficultyButtonsFragment was calling getString(), a
method which requires the fragment to be attached to an activity, in
onInflate(), which when the fragment is recreated is called when the
fragment is not attached to an activity, causing a crash. Fixed by
moving the getString() call to onAttach().