Pause in GameActivity in onPause()
GameActivity now pauses (i.e. brings up the pause overlay by calling
pause()) in onPause(), if it is not already paused and if the user has
not yet lost. Detecting lost status is done through a new boolean field,
hasLost, which is initially false and is set to true in onLose() and in
quit(). Thus, the game is paused whenever the user presses the home or
recent apps buttons, but not when they lose or quit the game.

As well, GameActivity no longer has android:noHistory="true" set in
AndroidManifest.xml; instead, finish() is now called in onLose() to
remove the activity. This means that the user can now return to the (now
paused) GameActivity after pressing the home or recent apps buttons.