Pause PostGameActivity animations in onPause()
PostGameActivity now keeps track of whether it has a new best score
(i.e. the new best score TextView is visible and pulsing) through the
boolean field isNewBestScore. If it is true, the field newBestScoreText
will be filled with @id/new_best_score_text; otherwise it is null.

These fields are used in onPause() and onResume() - if isNewBestScore is
true, onPause() will stop its animation and onResume() will start or
restart it.

Also, because onResume() is always called after onCreate(), onCreate()
no longer starts the animation with startAnimation(), but instead merely
sets it with setAnimation() and lets onResume() start it.