Save and restore GameActivity instance state
There are now two new methods in GameActivity: onSaveInstanceState()
(overridden) and restoreState() (new). onSaveInstanceState() saves the
state into the Bundle, and restoreState() restores it.

The keys saved into the Bundle are defined as constants in GameActivity.
They are:

- STATE_ROUND: (int) the `round` field
- STATE_SCORE: (int) the `score` field, and the contents of scoreText
- STATE_SCORE_FOR_ROUND: (int) the `scoreForRound` field
- STATE_VARIANT_ID: (int) the `variantId` field
- STATE_ALLOW_IMG_TAPS: (boolean) the `allowImgTaps` field
- STATE_HAS_LOST: (boolean) the `hasLost` field (possibly unnecessary)
- STATE_IS_PAUSED: (boolean) the `isPaused` field (key usually true)
- STATE_COUNTDOWN_CIRCLE_MAX: (int) the max of the countdown circle
- STATE_COUNTDOWN_CIRCLE_PROGRESS: (float) the progress of the circle
- STATE_COUNTDOWN_CIRCLE_VALUE: (int) the value displayed inside circle
- STATE_COUNTDOWN_CIRCLE_RESETTING: (boolean) the status of the circle's
  animation. A value of true means the reset animation is playing; false
  means the countdown animation is playing.
- STATE_IMG_PAIR_ID: (int) the ID of the current normal/variant pair.

From this, the state can be restored.

There are also two new methods in ImageSupplier: getPairId() and
getPairById(). The former returns the index of a given Pair<Drawable,
Drawable> in the ImageSupplier's list; the second gets a pair by said
index. The methods are used to keep track of which image pair is being
displayed.

Note that since onPause() nearly always calls pause(), the game will
nearly always restart paused. This is intended, but may not be obvious.