Add activity transitions
There are now four new, very simple anim files: slide_in_left.xml,
slide_in_right.xml, slide_out_left.xml, and slide_out_right.xml. The
animations are very similar: they are all linearly interpolated
translate animations with the same duration (the new resource
@integer/activity_transition_duration = 200), and none of them modify
the Y axis. The out animations modify the X axis from 0%p to either
100%p (slide_out_right) or -100% (slide_out_left), and the in animations
do the reverse: they modify it from either 100%p (slide_in_right) or
-100%p (slide_in_left) to 0%p.

These animations are used to transition from activity to activity. Util
has two new methods: getToLeftTransition() and getToRightTransition().
They both return Bundles that can be passed as the second argument to
startActivity() to specify the transition.

DifficultyButtonsFragment has a new XML attribute, transition_direction,
which is an enum with two possible values, "left" and "right". This
attribute specifies in which direction the activity should transition
when a difficulty button is clicked, and is necessary because the
fragment is used in places that want different transitions.

The transition directions combine to create the following effective
horizontal arrangement of the activities:

MainActivity | GameActivity | PostGameActivity

That is, when MainActivity transitions to GameActivity, it does so to
the right; GameActivity transitions to MainActivity to the left; and so
on, to create the effect. Of course, if the back button is pressed in
PostGameActivity, GameActivity will be skipped over to get to
MainActivity; this horizontal arrangement is thus imperfect, but the
effect is created anyways.