Add countdown before game
There is now a 3->2->1 countdown similar to the unpause countdown called
the starting countdown before the game.

Since the code for the starting and unpause countdowns turned out to be
*very* similar, the method Util.getCountdownAnimator() was added. It
takes a duration, a TextView to show the countdown, and a
Util.CountdownEndListener, a new interface with only an onEnd() method,
used to add code that runs at the end of the countdown.

The starting countdown's UI components consist of the countdown text,
@id/starting_countdown_text, and @id/starting_countdown_hint, the hint
text below it. The hint text is simply a string (@string/countdown_hint)
that explains what to do in the game. Eventually, the hint text may be
varied - before the first game, it may say a simple text, but it could
say other less obvious, but still helpful hints afterwards.

@id/starting_countdown_text is positioned in the centre of the screen,
but with a large offset (@dimen/starting_countdown_text_offset = 70dp).
@id/starting_countdown_hint is positioned below @id/starting_countdown_
text, with a large negative top margin (@dimen/starting_countdown_hint_
margin_top = -100dp). It should always be set to -@dimen/starting_
countdown_text_offset - 30dp.

The countdown is triggered by the new method startCountdownToGameStart()
which is called instead of startRound() when savedInstanceState == null.
When the countdown is finished, the new method hideStartingCountdown()
is called, then startRound().

hideStartingCountdown(), which is also called at the beginning of
restoreState(), hides the starting countdown text and hide and makes
visible the pause button, which is hidden at the beginning. This is
because it makes no sense to pause before the game has begun.