Generate main menu title text size
Before, the text size of the main menu title text (the one that says
"variant tap" currently) was hardcoded, and this would have to be
changed for every i18n and screen width. Now, Util.getLargestTextSize()
implements a binary search for the perfect text size.

The goal of Util.getLargestTextSize() is to find an integer text size
`size` such that `width(size) <= maxSize && width(size + 1) > maxSize`,
where `width` is a function giving the width of the text with a certain
text size and maxSize is the maximum width of the text.

Util.getLargestTextSize() takes a TextView, the text to go in it, a max
width in pixels, and a maximum font size (in sp). It performs a binary
search and returns the optimal font size as described above, or, if it
cannot find it, returns the maximum with a warning. It uses a new
private method, Util.measureText(), to get the width of the text for any
given font size.

Util.getLargestTextSize() is used in an OnGlobalLayoutListener in
MainActivity.onCreate(). It's used to set the font size of the title
TextView (@id/main_title) with the maximum width of the parent layout's
width minus twice @dimen/main_menu_title_side_padding = 20dp and the
maximum font size of @dimen/max_main_title_size = 150sp. An
OnGlobalLayoutListener is used so that the parent layout's width can be
gotten - if it was in the main onCreate() body, it would be 0 as the
layout would not yet have been laid out.