Make thrown images start/end offscreen, balance
There is now a FrameLayout in the middle of MainActivity called
@id/throwing_layout that spans the whole activity. The thrown images now
do their throwing using the top and left margins of it instead of the
vertical/horizontal biases of a ConstraintLayout (done through simply
multiplying the bias with the width/height). This way, the range of the
ValueAnimator can be adjusted to -0.2 to 1.2, with the images starting
and ending offscreen.

Also rebalance all the image throwing related resources.