Move circle buttons to below title, tweak them
The circle buttons have been moved from below the play button to below
the title. The play button has resumed its previous position, while the
top margin of the title has been adjusted from 50dp to 35dp. The circle
buttons' top margin stays the same, while their button margin has been
removed.

The image-throwing ranges have been adjusted to account for this: the
minimum vertical bias is now 0.35 (from 0.3), and the minimum vertical
bias difference is now 0.2 (from 0.25).

As well, the circle buttons are now transparent with a 2dp stroke, and
the colour of that stroke is now @color/colorPrimary.