Make first circle button a sound toggle
The first circle button in MainActivity is now a sound toggle. The
button has a music note in it (res/drawable/music_note.png) that, when
pressed, turns grayscale (@color/circleButtonDisabled) with a slash in
the circle. The button is currently nonfunctional.

The music note is stored completely white, and the colour changing is
accomplished through setting the drawable's filter. The circle with a
slash through it is an entirely separate drawable, stored in res/
drawable/circle_disabled.xml.

The toggling is accomplished through the new methods toggleSound(),
turnSoundButtonOff(), and turnSoundButtonOn() in MainActivity. The
toggled state is controlled by the boolean field soundOn.
turnSoundButton(On|Off)() change the colour filter of the music note,
the background drawable (circle or circle_disabled), and the content
description of the image (sound_on_desc or sound_off_desc).

The button is not functional, but will be soon.