Add about page for attribution
The "information" or "about" circle button (with an "i" in it) is now a
shortcut to the new AboutActivity, which contains a ScrollView-enclosed
vertical LinearLayout with many TextViews containing attributions and
such.

There are many new styles to style the TextViews consistently:
- AboutText is the base
- AboutText.Header is for a header (bigger, more margins, colours, etc.)
- AboutText.Body is for body text (small, black, standard margins)
- AboutText.Body.Key is for a key in a key-value pair (coloured)
- AboutText.Body.Value is for a value in a key-value pair
- AboutTextKeyValueLayout is for a LinearLayout with a key-value pair

All the strings used are in strings.xml.

In the metaphor produced by the transitions, AboutActivity is positioned
to the left of MainActivity. The user returns to MainActivity by
pressing the back button.