Add leaderboard icon, improve content descriptions
res/drawable/leaderboard.png is now a leaderboard icon, used in the
third circle button. Attribution is provided in the about screen.

Content descriptions have been improved: the sound off/on ones were
backwards, and "About this app" was unnecessarily verbose.

Also instead of "Code" it now says "Developer" on the about screen.