Add a losing animation
When the user loses the game (either by tapping a normal or running out
of time), an animation now plays. The animation is defined in the new
method getLoseAnimation() in GameActivity and is run in onLose(). It is
defined as an AnimatorSet containing a series of AnimatorSets and
ObjectAnimators. I'd love to define it in XML, but unfortunately there's
lots of dynamic dimensions, so that can't happen.

The animation
-------------

Stage 1: Normals fall.
Each normal falls to the bottom of the screen in @integer/lose_normal_
fall_time_ms (300) using an AccelerateInterpolator.

Stage 2: Variant moves to the centre and grows.
Wait @integer/lose_before_variant_time_ms (500). The variant moves from
its original position to the centre of the screen. At the same time, the
variant grows from its original dimensions to a width and length both of
@dimen/lose_variant_anim_size (175dp). Both take @integer/lose_variant_
move_time_ms (600) and use an AccelerateDecelerateInterpolator.

Stage 3: Variant pulses.
The variant pulses from @dimen/lose_variant_anim_size (175dp) to @dimen/
lose_variant_anim_pulse_size (185dp) over @integer/lose_variant_pulse_
time_ms (2500), pulsing @integer/lose_variant_pulse_times (3 times),
using a CycleInterpolator.

Stage 4: Variant falls.
The variant falls from its position in the centre of the screen to the
bottom of the screen in @integer/lose_variant_fall_time_ms (400) using
an AccelerateInterpolator. The screen advances to the PostGameActivity.

------------------------------------------------------------------------

Also, android:clipChildren and android:clipPadding are now false on both
the @id/imgs_grid and the root GameActivity layout to allow the images
to "fall" to the bottom of the screen.