Add raincloud normal and variant
The raincloud is composed of a cloud with a gradient from white at the
top to a shade of grey at the bottom, as well as a number of blue
raindrops below the cloud, with a shine on each of the raindrops. The
normal has a lighter shade of grey and five raindrops tilted with the
pointed top facing the right, while the variant has a darker shade of
grey and three raindrops tilted with the pointed top facing the left.