Add BaseGameUtils library as module
BaseGameUtils is a convenience library provided by Google for use with
the Google Play game services APIs. For some reason it's provided as
a subdirectory of the playgameservices/android-basic-samples repository
on GitHub instead of a repo in its own right, so it must be installed as
a module. It also uses out of date versions for its libraries.

BaseGameUtils' build.gradle checks several variables for existence, and
if it does not find them, it places an out-of-date version in them. The
variables are later used for library versions, causing an error. Thus,
the relevant variables are set in the top-level build.gradle, and for
good measure they're used in app/build.gradle too.

BaseGameUtils' build.gradle also hardcodes the buildToolsVersion -
this had to be replaced with the build_tools_version variable. Thus,
minimal changes were made to BaseGameUtils.

The BaseGameUtils module has been checked into version control only so
that you don't have to fetch it manually from GitHub in order to clone
the repository.

The only changes made to the BaseGameUtils module's code was to add
@SuppressWarnings("all") on all 5 classes to prevent obscene amounts of
warnings.