Add sign-in + showing leaderboard logic
MainActivity now uses GameHelper to automatically sign the user in, with
logic for what to do if the user signs out from within the default
leaderboard UI. The leaderboard circle button now works as well, and
brings up the aforementioned default leaderboard UI.

When the application starts, if the user has not canceled the sign-in
dialog at least once before, the automatic sign-in flow starts. If the
user cancels the dialog, they will not automatically be signed in again
until they sign in through the leaderboard button.

The leaderboard circle button acts as a sign-in button in that, if the
user is not signed in, it will initiate the sign-in flow as if they had
clicked a dedicated sign-in button. It will also allow the application
to automatically sign the user in on application start, even if they had
previously canceled it.

If the user signs out from within the leaderboard UI, MainActivity
catches this in onActivityResult(), disconnects the GameHelper, and
stops automatic sign-in as if they had signed out from a dedicated
sign-out button.

The different Google Play game services actions that can (currently) be
undertaken are enumerated in MainActivity.GPGSAction.

Also removed attribution for the old leaderboard icon which is no longer
in use.