Add GPGSHelperService, replace GameHelper with it
GPGSHelperService is a bound service. Its sole purpose is to encapsulate
a GoogleApiClient for multiple activities.

GPGSHelperService has a GoogleApiClient instance whose lifecycle it
maintains in sync with its own - it creates the client in onCreate() and
destroys it in onDestroy(). Its IBinder implementation -
GPGSHelperService.GPGSHelperBinder - simply returns the service itself.

GPGSHelperService has a number of public helper methods to interface
with the GoogleApiClient.
- onActivityResult() must be called in Activity.onActivityResult() in
  each Activity bound to the service.
- tryAutoConnect() attempts to automatically connect to GPGS, if the
  SharedPreferences preference Util.PREF_AUTO_CONNECT is true.
- connect() attempts to connect to GPGS.
- isConnected() returns whether the GoogleApiClient is connected.
- setActionOnSignIn() sets a GPGSAction to be performed on sign-in.
- getApiClient() gets the GoogleApiClient.

Whether to automatically connect in tryAutoConnect() is determined by
the Util.PREF_AUTO_CONNECT SharedPreference. The pref is modified when
the user successfully signs in, cancels the sign-in flow, or signs out
via a default GPGS UI.

The new class GPGSHelperServiceConnection eases in binding to
GPGSHelperService. To bind, implement GPGSHelperServiceConnection
.ServiceReceiver in your Activity and store the GPGSHelperService that
will be passed to the receiveService() method in the ServiceReceiver.

GameHelper has been completely removed from the app code and has been
replaced with GPGSHelperService, which fulfills practically the same
function but allows for multiple activities.