Add sign-in/sign-out button in AboutActivity
At the bottom of AboutActivity, there is now a button that toggles
between signing in and signing out. Its text is "Sign [in|out] of Google
Play" (@string/gpgs_sign_[in|out]) with the GPGS controller icon
(@drawable/controller) at the end of the text. The text and icon are
together centered in the button; this is taken care of programatically
in AboutActivity.centerButtonTextAndIcon(). In order to scale it down to
a reasonable size, the icon is added in AboutActivity.onCreate(). The
icon has @dimen/gpgs_icon_padding (10dp) of padding at the top and
bottom of the button and the icon and text have @dimen/gpgs_icon_spacing
(10dp) of spacing between them.

AboutActivity now binds to GPGSHelperService in order to implement this
button. When the button is pressed, signInOrOutOfGPGS() is called, a
new method which either signs in or out of GPGS depending on whether it
is signed in.

Several methods were added and changes were made in GPGSHelperService to
facilitate this:
- new method: signOut() signs out of GPGS and disconnects the client.
- new method: connectWithoutSignInFlow() attempts to connect, but will
  not attempt to solve any connection failures (and thus will not invoke
  the sign-in flow). It should be called in the receiveService() methods
  of any activities binding to GPGSHelperService which do not call
  tryAutoConnect(). It calls connect(activity, false).
- new private method: connect(Activity, boolean trySignIn) is the same
  as the old connect(Activity), except it sets trySignIn to the
  parameter of the same name.
- connect(Activity) now merely calls connect(activity, true).
- Many warnings have been removed because they have perfectly legitimate
  use-cases.

As well, there is a new GPGSAction: HideSignInButton. This GPGSAction is
used in AboutActivity. It checks whether the activity implements the new
interface HasSignInButton, then calls its hideSignInButton() method (the
only method in the interface).

Also, @style/ButtonStyle has been changed - button are no longer in all
caps. This is because it looked weird with the sign in/out button.