Refactor circle buttons into custom views
Before, all the circle buttons were ImageButtons with all the control
code in MainActivity. That was not modular!

Now the circle buttons are a class hierarchy of custom views.
BaseCircleButton is the base and each circle button (AboutCircleButton,
LeaderboardCircleButton, RemoveAdsCircleButton, and SoundToggleCircle-
Button) inherits from it. All the styles are defined in styles.xml and
attrs.xml.

Also move GPGSHelperServiceConnection.ServiceReceiver into its own top-
level interface, GPGSHelperServiceClient, which all clients of
GPGSHelperService must implement. It has one method more than Service-
Receiver, getService(), which returns the service previously received.
It's used in LeaderboardCircleButton to get a GPGSHelperService.

The modular circle buttons will in the future be added to other views
such as the pause overlay and PostGameActivity.