Add circle buttons to the pause overlay
The sound toggle and remove ads buttons have been added to the pause
overlay.

BaseCircleButton has been modified to have a "baseColor" attribute that
gives the "base color" of the icon and circle. It and the sound toggle
button manage the colors of the circle buttons so that the circle
buttons in the pause overlay have the correct colors (@color/gamePaused-
Main).

Also MainActivity no longer handles anything about circle buttons or
their colour.