Implement achievement 'Mesmerizing...'
The achievement is unlocked by not moving from MainActivity for two
minutes (presumably staring at the animation). This is implemented
through a dedicated Timer in MainActivity which is started in onResume()
and cancelled in onPause(). The amount of time to wait in milliseconds
is handled as @integer/achievement_stare_animation_time_ms.