Fix bug with images being stretched out sometimes
The GridLayout solution in GameActivity for storing the grid of images
was massively bugged. On some screens (e.g. the Nexus S 480x800 screen),
the images were super stretched out across the whole screen and the game
was unplayable.

This was fixed by completely removing GridLayout and changing to a
ConstraintLayout. The ImageSwitchers are no longer dynamically
generated and added to the layout; instead, there are three layouts,
@layout/grid_2x2, @layout/grid_3x2, and @layout/grid_3x3, which are
merged into the ConstraintLayout.

As well, ImageSupplier has been changed to load the images with proper
intrinsic widths and heights, and GameActivity's root layout has been
changed to ConstraintLayout from RelativeLayout.

Resolves #33.