Fix the main menu title on tablets
Util.getLargestTextSize() was going into an infinite loop when the
target/maximum size in pixels was too large. Now it does not. This was
causing the app to fail to load on tablets where the width was too
large.

Also, the maximum font size of the main menu title text was too small;
it is now larger.

As well, the title text looked kinda stupid when it was absolutely
enormous on a tablet screen; on tablets in portrait mode, the margins
on each side are now significantly bigger.